api/requirements.txt
flask
redis
pymongo
pyyaml
